# official/api.openapi.upgraded.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vigor - Gym Management API",
    "version": "1.0.0",
    "description": "AI-powered gym management platform for Mexico"
  },
  "servers": [
    {"url": "https://api.vigor.mx/v1", "description": "Production"},
    {"url": "https://staging-api.vigor.mx/v1", "description": "Staging"}
  ],
  "security": [{"bearerAuth": []}],
  "paths": {
...
    "/plans": {
      "get": {
        "operationId": "listPlans",
        "summary": "List membership plans",
        "tags": ["Plans"],
        "parameters": [
          {"name": "gymId", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/Plan"}
                }
              }
            }
          }
        },
        "x-roles": ["owner", "manager", "staff", "member"],
        "x-entity": "Plan"
      },
      "post": {
        "operationId": "createPlan",
        "summary": "Create membership plan",
        "tags": ["Plans"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreatePlanInput"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Plan"}
              }
            }
          }
        },
        "x-roles": ["owner", "manager"],
        "x-entity": "Plan"
      }
    },
...
    "/members": {
      "get": {
        "operationId": "listMembers",
        "summary": "List members",
        "tags": ["Members"],
        "parameters": [
          {"name": "gymId", "in": "query", "required": true, "schema": {"type": "string", "format": "uuid"}},
          {"name": "status", "in": "query", "schema": {"$ref": "#/components/schemas/MemberStatus"}},
          {"name": "churnRisk", "in": "query", "schema": {"$ref": "#/components/schemas/ChurnRisk"}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 50}},
          {"name": "offset", "in": "query", "schema": {"type": "integer", "default": 0}}
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Member"}
                    },
                    "total": {"type": "integer"},
                    "limit": {"type": "integer"},
                    "offset": {"type": "integer"}
                  }
                }
              }
            }
          }
        },
        "x-roles": ["owner", "manager", "staff"],
        "x-entity": "Member"
      },
      "post": {
        "operationId": "createMember",
        "summary": "Create member",
        "tags": ["Members"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateMemberInput"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Member"}
              }
            }
          }
        },
        "x-roles": ["owner", "manager", "staff"],
        "x-entity": "Member"
      }
    },
    "/members/{id}": {
      "get": {
        "operationId": "getMember",
        "summary": "Get member",
        "tags": ["Members"],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Member"}
              }
            }
          },
          "404": {"$ref": "#/components/responses/NotFound"}
        },
        "x-roles": ["owner", "manager", "staff", "member"],
        "x-entity": "Member"
      },
      "patch": {
        "operationId": "updateMember",
        "summary": "Update member",
        "tags": ["Members"],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/UpdateMemberInput"}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Member"}
              }
            }
          }
        },
        "x-roles": ["owner", "manager", "staff", "member"],
        "x-entity": "Member"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"}
          }
        }
      },
...
    },
    "schemas": {
      "UUID": {"type": "string", "format": "uuid"},
      "ISODate": {"type": "string", "format": "date-time"},
      "Date": {"type": "string", "format": "date"},
      "MemberStatus": {"type": "string", "enum": ["active", "inactive", "suspended", "cancelled"]},
...
      "ChurnRisk": {"type": "string", "enum": ["low", "medium", "high"]},
      "BillingCycle": {"type": "string", "enum": ["monthly", "quarterly", "annually", "lifetime"]},
...
      "Plan": {
        "type": "object",
        "properties": {
          "id": {"$ref": "#/components/schemas/UUID"},
          "gymId": {"$ref": "#/components/schemas/UUID"},
          "name": {"type": "string"},
          "description": {"type": "string", "nullable": true},
          "priceMXN": {"type": "integer"},
          "billingCycle": {"$ref": "#/components/schemas/BillingCycle"},
          "trialDays": {"type": "integer"},
          "features": {"type": "object", "nullable": true},
          "isActive": {"type": "boolean"},
          "createdAt": {"$ref": "#/components/schemas/ISODate"}
        },
        "required": ["id", "gymId", "name", "priceMXN", "billingCycle", "trialDays", "isActive", "createdAt"]
      },
      "CreatePlanInput": {
        "type": "object",
        "properties": {
          "gymId": {"$ref": "#/components/schemas/UUID"},
          "name": {"type": "string"},
          "description": {"type": "string"},
          "priceMXN": {"type": "integer", "minimum": 0},
          "billingCycle": {"$ref": "#/components/schemas/BillingCycle"},
          "trialDays": {"type": "integer", "minimum": 0, "default": 0},
          "features": {"type": "object"}
        },
        "required": ["gymId", "name", "priceMXN", "billingCycle"]
      },
      "Member": {
        "type": "object",
        "properties": {
          "id": {"$ref": "#/components/schemas/UUID"},
          "gymId": {"$ref": "#/components/schemas/UUID"},
          "locationId": {"$ref": "#/components/schemas/UUID"},
          "email": {"type": "string", "format": "email"},
          "fullName": {"type": "string"},
          "phone": {"type": "string", "nullable": true},
          "birthdate": {"$ref": "#/components/schemas/Date"},
          "gender": {"type": "string", "enum": ["male", "female", "other", "prefer_not_to_say"], "nullable": true},
          "emergencyContact": {"type": "object", "nullable": true},
          "consentBiometric": {"type": "boolean"},
          "consentMarketing": {"type": "boolean"},
          "status": {"$ref": "#/components/schemas/MemberStatus"},
          "joinedAt": {"$ref": "#/components/schemas/ISODate"},
          "lastVisitAt": {"$ref": "#/components/schemas/ISODate"},
          "churnScore": {"type": "number", "nullable": true},
          "churnRisk": {"$ref": "#/components/schemas/ChurnRisk"},
          "createdAt": {"$ref": "#/components/schemas/ISODate"},
          "updatedAt": {"$ref": "#/components/schemas/ISODate"}
        },
        "required": ["id", "gymId", "email", "fullName", "consentBiometric", "consentMarketing", "status", "joinedAt", "createdAt", "updatedAt"]
      },
      "CreateMemberInput": {
        "type": "object",
        "properties": {
          "gymId": {"$ref": "#/components/schemas/UUID"},
          "locationId": {"$ref": "#/components/schemas/UUID"},
          "email": {"type": "string", "format": "email"},
          "fullName": {"type": "string"},
          "phone": {"type": "string"},
          "birthdate": {"$ref": "#/components/schemas/Date"},
          "gender": {"type": "string", "enum": ["male", "female", "other", "prefer_not_to_say"]},
          "emergencyContact": {"type": "object"},
          "consentBiometric": {"type": "boolean", "default": false},
          "consentMarketing": {"type": "boolean", "default": false}
        },
        "required": ["gymId", "email", "fullName"]
      },
      "UpdateMemberInput": {
        "type": "object",
        "properties": {
          "email": {"type": "string", "format": "email"},
          "fullName": {"type": "string"},
          "phone": {"type": "string"},
          "birthdate": {"$ref": "#/components/schemas/Date"},
          "gender": {"type": "string", "enum": ["male", "female", "other", "prefer_not_to_say"]},
          "emergencyContact": {"type": "object"},
          "consentBiometric": {"type": "boolean"},
          "consentMarketing": {"type": "boolean"},
          "status": {"$ref": "#/components/schemas/MemberStatus"}
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {"type": "string"},
          "code": {"type": "string"},
          "details": {"type": "object"}
        },
        "required": ["message"]
      }
    }
  }
}